requirements.txt
elasticsearch
python-dotenv
tqdm